+++ system-config.js
@@ -1,20 +1,50 @@
-// backup original SystemJS methods
+// Backup original SystemJS methods
 var _System = {
     normalize: System.normalize,
+    normalizeSync: System.normalizeSync,
     locate: System.locate,
     fetch: System.fetch,
     translate: System.translate,
     instantiate: System.instantiate
 };
 
-System.baseURL = 'meteor://';
+// Make `register` the default module format
 System.config({
     meta: {
-        'meteor://*': {
+        '*': {
             format: 'register'
         }
     }
 });
+
+// Regular expressions for Meteor package import syntax
+var appRegex = /^\{}\//;
+var packageRegex = /^{([\w-]*?):?([\w-]+)}/;
+var packageRegexBC = /^([\w-]+):([\w-]+)/;
+
+/**
+ * Convert Meteor package syntax to System.normalize friendly string.
+ * The `__author_package/foo` syntax in an internal implementation that is subject to change.
+ * You should never rely on it! Instead pass all your module names through System.normalizeSync
+ * @param {string} name - unnormalized module name with Meteor package syntax
+ * @returns {string} - unnormalized module name without Meteor package syntax
+ */
+var normalizeMeteorPackageName = function (name) {
+    name = name
+        .replace(appRegex, '') // {}/foo -> foo
+        .replace(packageRegex, '__$1_$2'); // {author:package}/foo -> __author_package/foo
+
+    if(packageRegexBC.test(name)){
+        // provide temporary backward compatibility for versions < 0.4 package syntax
+        console.warn([
+            '[Universe Modules]',
+            'You are using deprecated syntax for importing modules from a package.',
+            'Instead of', name, 'you should use', name.replace(packageRegexBC, '{$1:$2}')
+        ].join(' '));
+        return name.replace(packageRegexBC, '__$1_$2'); // author:package/foo -> __author_package/foo
+    }
+    return name;
+};
 
 /*
  * name: the unnormalized module name
@@ -22,18 +52,22 @@ System.config({
  * parentAddress: the address of the requesting module
  */
 System.normalize = function (name, parentName, parentAddress) {
-    // add meteor prefix
-    if (name[0] !== '.' && name[0] !== '/' && name.indexOf('://') === -1) {
-        name = 'meteor://' + name;
-    }
 
-    // allow foomodule.import syntax in import name
+    // Allow foomodule.import syntax in import name (TypeScript support)
     if (name.slice(-7) === '.import') {
         name = name.slice(0, -7);
     }
 
-    // load original normalize
-    return _System.normalize.call(this, name, parentName, parentAddress);
+    // Load original normalize
+    return _System.normalize.call(this, normalizeMeteorPackageName(name), parentName, parentAddress);
+};
+
+/*
+ * name: the unnormalized module name
+ * parentName: the canonical module name for the requesting module
+ */
+System.normalizeSync = function (name, parentName) {
+    return _System.normalizeSync.call(this, normalizeMeteorPackageName(name), parentName);
 };
 
 /*
@@ -59,14 +93,9 @@ System.fetch = function (load) {
         return promise;
     }
 
-    if(load.name.slice(0, 9) !== 'meteor://'){
-        // not our protocol, ignore
-        return promise;
-    }
-
-    // show our warning
+    // Add our warning
     return promise.catch(function (err) {
-        console.warn('[Universe Modules]: Module ' + load.name.slice(9) + ' does not exist! You will probably see other errors in the console because of that.');
+        console.warn('[Universe Modules]: Module ' + load.name.replace(System.baseURL, '') + ' does not exist! You will probably see other errors in the console because of that.');
     });
 };