+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 
+## 0.4
+
+- Introduce new syntax `{author:package}/foo` for importing modules from packages (compatible with MDG less/stylus import syntax)
+- Provide backward compatibility with warning message for previous syntax
+- Better handling of SystemJS module normalization
+- Support for System map/packages api with `System.normalizeSync`
+- Remove `meteor://` protocol
+- Update `babel-compiler` to 5.8.3_1 
+- Update `babel-runtime` to 0.1.2
+
 ## 0.3
 
 - Update SystemJS to 0.18.4